hooked up real time socket messaging to database
