+++ app/(routes)/channel/[id]/page.tsx
@@ -1,23 +1,7 @@
 import { authOptions } from '@/utils/auth';
 import { getServerSession } from 'next-auth';
 import { prisma } from '@/utils/prisma';
-import ChannelInput from '@/components/channel/channelInput';
-import Messages from '@/components/channel/messages';
-
-type user = {
-  username: string,
-}
-
-type message = {
-  content: string,
-  user: user,
-  sentAt: Date,
-};
-
-type channel = {
-  messages: message[],
-  name: string,
-};
+import ChatWindow from '@/components/channel/chatWindow';
 
 // TODO setup proper channels (private and public)
 export default async function Page({ params }: { params: { id: string } }) {
@@ -47,18 +31,10 @@ export default async function Page({ params }: { params: { id: string } }) {
   if (session?.user) {
     return (
       <div className='flex min-h-screen flex-col items-center justify-between p-24'>
-        Your username is: {session.user.username} and we are on channel: {params.id}
-        <Messages user={session.user.username} channelName='Test' />
-        <div>
-          {channel?.messages.map((element: message, index: number) => (
-            <div className='mb-8' key={index}>
-              <p>sent by: {element.user.username}</p>
-              <p>content: {element.content}</p>
-              <p>Sent At: {element.sentAt.toString()}</p>
-            </div>
-          ))}
-        </div>
-        <ChannelInput user={session.user.username} channelName='Test' />
+        <p className='font-bold text-white text-xl'>
+          Your username is: {session.user.username} and we are on channel: {params.id}
+        </p>
+        <ChatWindow user={session.user.username} channelName='Test' currentMessages={channel?.messages} />
       </div>
     );
   }
